backend/requirements.txt
asgiref==3.5.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.7.1
Django==4.1.4
django-colorfield==0.8.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.7
